dynamic select options in pop up modal

+++ server/routes/collections.router.js
@@ -2,6 +2,7 @@ const express = require('express');
 const router = express.Router();
 const pool = require('../modules/pool.js');
 
+//inner gets the movies inside the lists
 router.post('/', (req,res)=>{    
     let queryString = `INSERT INTO "movie_list" ("list_id", "movie_id") VALUES ($1,$2)`;
     pool.query(queryString, [req.body.data.list, req.body.data.movie]).then((results) => {        
@@ -11,6 +12,7 @@ router.post('/', (req,res)=>{
         res.sendStatus(500);
     })
 })
+//outer gets names of lists created per user
 router.post('/outer', (req,res)=>{      
     let queryString = `INSERT INTO "user_lists" ("list_title", "user_id") VALUES $1, $2;`;
     pool.query(queryString, [req.body.newListTitle, req.session.passport.user]).then((results) => {        
@@ -24,28 +26,28 @@ router.get('/outer', (req,res)=>{
     console.log('in List OUTER with req ', req.session.passport.user);
     
     let queryString = `SELECT "list_title","user_lists"."id" FROM "user_lists"
-    JOIN "movie_list" ON "movie_list"."list_id" = "user_lists"."id"
+    LEFT JOIN "movie_list" ON "movie_list"."list_id" = "user_lists"."id"
     JOIN "user" ON "user"."id" = "user_lists"."user_id"
     WHERE "user"."id" = $1
     GROUP BY "list_title", "user_lists"."id";`
-    pool.query(queryString, req.session.passport.user).then((results) => {
-        // console.log('what is results.rows here? ', results.rows);
+    pool.query(queryString, [req.session.passport.user]).then((results) => {
+        console.log('what is results.rows here? ', results.rows);
         res.send(results.rows);
     }).catch((err) => {
         console.log(err);
         res.sendStatus(500);
     })
 })
-router.delete('/:listId', (req,res)=>{
-    let queryString = `DELETE FROM "user_lists" WHERE "id" = $1;`;
-    pool.query(queryString, [req.params.listId]).then((results) => {        
+router.delete('/outer/:listId', (req,res)=>{
+    let queryString = `DELETE FROM "user_lists" WHERE "id" = $1 AND "user_lists"."user_id" = $2;`;
+    pool.query(queryString, [req.params.listId, req.session.passport.user]).then((results) => {        
         res.sendStatus(200);
     }).catch((err) => {
         console.log(err);
         res.sendStatus(500);
     })
 })
-router.put('/:listId', (req,res)=>{
+router.put('/outer/:listId', (req,res)=>{
     let queryString = `UPDATE "user_lists" SET "list_title"='$1' WHERE "id" = $2;`;
     pool.query(queryString, [req.body,req.params.listId]).then((results) => {        
         res.sendStatus(200);
